requirements.txt
# 核心框架
streamlit>=1.30.0
pandas>=2.1.4
numpy>=1.26.2
joblib>=1.3.2

# 机器学习
scikit-learn==1.3.2
umap-learn>=0.5.5
shap>=0.42.1
scipy>=1.11.4

# 可视化
matplotlib>=3.8.2
seaborn>=0.13.2
kmapper>=2.0.1
networkx>=3.2.1

# 其他依赖
pillow>=10.1.0
rlPyCairo>=0.2.0
pycairo>=1.25.1